Remove Unused QuestionPayload Model #150
